修改user属性id的类别，mapper文件增加新的方法，.yml增加包下面的别名设置

+++ src/main/java/com/chinaopensource/apiserver/system/user/mapper/UserMapper.java
@@ -39,6 +39,6 @@ public interface UserMapper {
 	 * @param status
 	 * @return
 	 */
-	Boolean updateStatus(Long id,Integer status);
+	Boolean updateStatus(Integer id,Integer status);
 
 }
